refactor: rename `datamodule` to `module_data` and `module` to `module_model`
- Updated import paths across codebase and tests to reflect new module names.
- Ensured consistent references in documentation, comments, and test fixtures.
- Addressed all affected helper functions and regression tests.
