+++ src/rfdetr/training/__init__.py
@@ -25,8 +25,8 @@ from rfdetr.training.callbacks import (
 )
 from rfdetr.training.checkpoint import convert_legacy_checkpoint
 from rfdetr.training.cli import RFDETRCli
-from rfdetr.training.datamodule import RFDETRDataModule
-from rfdetr.training.module import RFDETRModelModule
+from rfdetr.training.module_data import RFDETRDataModule
+from rfdetr.training.module_model import RFDETRModelModule
 from rfdetr.training.trainer import build_trainer
 from rfdetr.utilities.logger import get_logger
