+++ src/rfdetr/training/module_model.py
