+++ tests/cli/test_smoke.py
@@ -45,8 +45,8 @@ ALL_CONFIGS = [
 def _run_cli(*args: str) -> int:
     """Run RFDETRCli in-process with the given args; return the SystemExit code."""
     from rfdetr.training.cli import RFDETRCli
-    from rfdetr.training.datamodule import RFDETRDataModule
-    from rfdetr.training.module import RFDETRModelModule
+    from rfdetr.training.module_data import RFDETRDataModule
+    from rfdetr.training.module_model import RFDETRModelModule
 
     with pytest.raises(SystemExit) as exc_info:
         RFDETRCli(RFDETRModelModule, RFDETRDataModule, args=list(args))
